Fix Scala compilation error - convert Int to Long for type compatibility
Error: "cannot be applied to (org.scalatest.matchers.Matcher[Int])" at line 313

Root cause: metricsPublisher.getRecordsWrittenSnapshot() returns Long, but (2 * items.size) is Int.
The matcher `be >= (2 * items.size)` creates Matcher[Int], causing type mismatch when applied to Long.

Fix: Convert comparison value to Long with .toLong

+++ sdk/cosmos/azure-cosmos-spark_3/src/test/scala/com/azure/cosmos/spark/PointWriterITest.scala
@@ -310,7 +310,7 @@ class PointWriterITest extends IntegrationSpec with CosmosClient with AutoCleana
     // This prevents race conditions where metrics snapshot is taken before all writes are recorded
     // Use eventually block to poll until the expected count is reached
     eventually(timeout(10.seconds), interval(100.milliseconds)) {
-      metricsPublisher.getRecordsWrittenSnapshot() should be >= (2 * items.size)
+      metricsPublisher.getRecordsWrittenSnapshot() should be >= (2 * items.size).toLong
     }
 
     metricsPublisher.getRecordsWrittenSnapshot() shouldEqual 2 * items.size
